refactor: models 폴더 domains로 변경
